refactor: merged the search component tests into one file
